% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_backup_files.R
\name{SR_backup_files}
\alias{SR_backup_files}
\title{Backup files}
\usage{
SR_backup_files(
  path_from = NULL,
  path_to = NULL,
  include = NULL,
  exclude = NULL,
  return_lists = FALSE
)
}
\arguments{
\item{path_from}{character: path to folder and files as backup input}

\item{path_to}{character: path to folder and files as backup output}

\item{include}{character: files to be included from backup (partial string match),
all other files will be excluded}

\item{exclude}{character: files to be excluded from backup (partial string match)}

\item{return_lists}{boolean: return lists of files from the backup process}
}
\value{
NULL, if 'return_lists = FALSE' then list
}
\description{
Function to backup files of a specific directory
}
\examples{
SR_backup_files(path_from = getwd(),
                path_to = paste0(getwd(), "/my_backup_folder/"),
                include = c("license"),
                exclude = c("backup"))

}
