% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_correlation_plot.R
\name{SR_correlation_plot}
\alias{SR_correlation_plot}
\title{Correlation plot}
\usage{
SR_correlation_plot(
  df,
  cor_threshold = NULL,
  save = FALSE,
  filename = "NA_Correlation.png",
  path_output = NULL
)
}
\arguments{
\item{df}{data.frame}

\item{cor_threshold}{numeric between 0 and 1}

\item{save}{Boolean}

\item{filename}{character, must end with "*png"}

\item{path_output}{character}
}
\value{
matrix, prints a plot
}
\description{
Plot correlation of all numeric, integer and date variables of a data.frame.
}
\examples{
data("mtcars")
SR_correlation_plot(df = mtcars)

}
