% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_difftime_years.R
\name{SR_difftime_years}
\alias{SR_difftime_years}
\title{Function to calculate correct years with floating points between two dates}
\usage{
SR_difftime_years(enddate, startdate)
}
\arguments{
\item{enddate}{a vector of class Date representing the end date}

\item{startdate}{a vector of class Date representing the start date}
}
\value{
A numeric vector of ages the same length as the startdate vector
}
\description{
Sources:
- https://github.com/jknowles/eeptools/blob/main/R/age_calc.R
- https://stackoverflow.com/a/78022697/4786677
}
\examples{
SR_difftime_years(enddate = "2024-11-30", startdate = "2023-11-29")
SR_difftime_years(enddate = "2023-11-30", startdate = "2022-11-29")

}
