% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_end_of_month.R
\name{SR_end_of_month}
\alias{SR_end_of_month}
\title{Calculate date at end of month}
\usage{
SR_end_of_month(date)
}
\arguments{
\item{date}{date}
}
\value{
date
}
\description{
by adding a month, then subtract a day
}
\examples{
SR_end_of_month("2019-12-24")

}
