% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_feat_eng_date.R
\name{SR_feat_eng_date}
\alias{SR_feat_eng_date}
\title{Feature engineering for variables of class "Date"}
\usage{
SR_feat_eng_date(df, only_date_to_numeric = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{only_date_to_numeric}{boolean (default = TRUE): if FALSE then in addition
the month, day and weekday of date-variables are added as new columns}
}
\value{
Boolean value TRUE or FALSE
}
\description{
Lots of common feature engineering utilities. Class "Date" is determined by
function SR_is_date()
}
\examples{
df <- data.frame(var_numeric = c(1, 2, 2, 3, NA, 4),
                 var_character = c("a", NA, "b", "c", "c", "d"),
                 var_factor = factor(c("a", "b", "c", "c", "d", NA)),
                 var_date = seq.Date(as.Date("2020-04-12"), by = "day", length.out = 6),
                 stringsAsFactors = FALSE)
SR_feat_eng_date(df)
SR_feat_eng_date(df, only_date_to_numeric = FALSE)
rm(df)

}
