% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_feat_eng_factors.R
\name{SR_feat_eng_factors}
\alias{SR_feat_eng_factors}
\title{Feature engineering for variables of class "factor" and "character"}
\usage{
SR_feat_eng_factors(
  df,
  order_weekday_month = TRUE,
  make_na_explicit = TRUE,
  replace_na_by_modus = FALSE,
  label_encoding = TRUE,
  count_encoding = FALSE,
  rank_encoding = FALSE,
  weighted_effect_coding = FALSE,
  target_encoding = FALSE,
  polynomial_encoding = FALSE,
  combine_rare_levels = TRUE,
  prop = 0.01,
  folds_index = NULL,
  use_other_df = NULL
)
}
\arguments{
\item{df}{data.frame}

\item{order_weekday_month}{boolean (default = TRUE): put weekdays and month in
natural order (factor levels, german and english)}

\item{make_na_explicit}{boolean (default = TRUE): na_level = "Missing"}

\item{replace_na_by_modus}{boolean (default = FALSE): alternative NA treatment}

\item{label_encoding}{boolean (default = TRUE): not implemented atm}

\item{count_encoding}{boolean (default = FALSE): not implemented atm}

\item{rank_encoding}{boolean (default = FALSE): not implemented atm}

\item{weighted_effect_coding}{boolean (default = FALSE): not implemented atm}

\item{target_encoding}{boolean (default = FALSE): not implemented atm}

\item{polynomial_encoding}{boolean (default = FALSE)}

\item{combine_rare_levels}{boolean (default = TRUE)}

\item{prop}{boolean (default = 0.01): required if 'combine_rare_levels = TRUE'}

\item{folds_index}{boolean (default = NULL):}

\item{use_other_df}{character (default = NULL): use other provided data.frame
for calculating statistics}
}
\value{
data.frame
}
\description{
Lots of common feature engineering utilities.
}
\examples{
df <- data.frame(var_numeric = c(1, 2, 2, 3, NA, 4),
                 var_character = c("a", NA, "b", "c", "c", "d"),
                 var_factor = factor(c("a", "b", "c", "c", "d", NA)),
                 var_date = seq.Date(as.Date("2020-04-12"), by = "day", length.out = 6),
                 stringsAsFactors = FALSE)
SR_feat_eng_date(df)
SR_feat_eng_date(df, only_date_to_numeric = FALSE)
rm(df)

}
