% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_feat_eng_numeric.R
\name{SR_feat_eng_numeric}
\alias{SR_feat_eng_numeric}
\title{Feature engineering for variables of class "numeric"}
\usage{
SR_feat_eng_numeric(
  df,
  trim_outliers = FALSE,
  replace_NA_special_value = FALSE,
  replace_NA_median = FALSE,
  log_scale_p1 = FALSE,
  interactions = FALSE,
  folds_index = NULL,
  exception = NULL,
  use_other_df = NULL
)
}
\arguments{
\item{df}{data.frame}

\item{trim_outliers}{boolean (default = FALSE): limit = 5 * standard deviation}

\item{replace_NA_special_value}{boolean (default = FALSE): trying to find a special
value like -1, 1, or 999*}

\item{replace_NA_median}{boolean (default = FALSE): alternative NA treatment,
creates a separate column to flag NA existence}

\item{log_scale_p1}{boolean (default = FALSE): log_scale, i.e. if variable is skewed}

\item{interactions}{boolean (default = FALSE): calculate variables with interactions
with "_plus_", "_multi_", "_minus_", "_div_"}

\item{folds_index}{boolean (default = NULL): not implemented atm}

\item{exception}{character (default = NULL): except variables from feature engineering}

\item{use_other_df}{character (default = NULL): use other provided data.frame
for calculating statistics}
}
\value{
data.frame
}
\description{
Lots of common feature engineering utilities.
}
\examples{
df <- data.frame(var_numeric = c(1, 2, 2, 3, NA, 4),
                 var_character = c("a", NA, "b", "c", "c", "d"),
                 var_factor = factor(c("a", "b", "c", "c", "d", NA)),
                 var_date = seq.Date(as.Date("2020-04-12"), by = "day", length.out = 6),
                 stringsAsFactors = FALSE)
SR_feat_eng_date(df)
SR_feat_eng_date(df, only_date_to_numeric = FALSE)
rm(df)

}
