% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_feat_eng_rows.R
\name{SR_feat_eng_rows}
\alias{SR_feat_eng_rows}
\title{Feature engineering for rows}
\usage{
SR_feat_eng_rows(
  df,
  row_count_na = TRUE,
  row_count_zero = TRUE,
  row_count_negative = TRUE,
  row_min = TRUE,
  row_mean = TRUE,
  row_max = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{row_count_na}{boolean (default = TRUE): all variables}

\item{row_count_zero}{boolean (default = TRUE): numeric values if > 1}

\item{row_count_negative}{boolean (default = TRUE): numeric values if > 1}

\item{row_min}{boolean (default = TRUE): numeric values if > 1}

\item{row_mean}{boolean (default = TRUE): numeric values if > 1}

\item{row_max}{boolean (default = TRUE): numeric values if > 1}
}
\value{
data.frame
}
\description{
Lots of common feature engineering utilities. Currently most of the KPIs support
only numeric variables.
}
\examples{
df <- data.frame(var_numeric1 = c(-1, 0, 2, 3, NA, 4),
                 var_numeric2 = c(-5, 0, -2, NA, 3, 4),
                 var_character = c("a", NA, "b", "c", "c", "d"),
                 var_factor = factor(c("a", "b", "c", "c", "d", NA)),
                 var_date = seq.Date(as.Date("2020-04-12"), by = "day", length.out = 6),
                 stringsAsFactors = FALSE)
SR_feat_eng_rows(df)
rm(df)

}
