% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_rstudio_jobs.R
\name{SR_get_rstudio_jobs}
\alias{SR_get_rstudio_jobs}
\title{Return a tibble of Jobs shown in RStudio's Jobs pane}
\usage{
SR_get_rstudio_jobs()
}
\value{
tibble
}
\description{
This is a minimal wrapper that fetches the job list and maps a few
common fields into columns. It requires running inside RStudio.
}
\examples{
\dontrun{
  SR_get_rstudio_jobs()
}
}
