% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_is_date.R
\name{SR_is_date}
\alias{SR_is_date}
\title{Check if object is class date}
\usage{
SR_is_date(x)
}
\arguments{
\item{x}{object}
}
\value{
Boolean value TRUE or FALSE
}
\description{
because missing function is.Date in base R
}
\examples{
SR_is_date(Sys.Date())

}
