% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_is_number.R
\name{SR_is_number}
\alias{SR_is_number}
\title{Check if object is a number}
\usage{
SR_is_number(x)
}
\arguments{
\item{x}{object}
}
\value{
Boolean value TRUE or FALSE
}
\description{
check if object is.numeric(), is.integer() or SR_is_date()
}
\examples{
SR_is_number(42.2)

}
