% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_join_check.R
\name{SR_join_check}
\alias{SR_join_check}
\title{Join-Check}
\usage{
SR_join_check(LHS = NULL, RHS = NULL)
}
\arguments{
\item{LHS}{data.frame}

\item{RHS}{data.frame}
}
\value{
None, results will be printed
}
\description{
Check the number of rows after different joining methods of two data.frames.
The joins will be done by all matching columns (identical names in both data.frames).
}
\examples{
SR_join_check(data.frame(a = 1:3),
              data.frame(a = c(3, 3:5)))

}
