% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_labeller_function.R
\name{SR_labeller_function}
\alias{SR_labeller_function}
\title{Add labels to a data.frame}
\usage{
SR_labeller_function(df, labels)
}
\arguments{
\item{df}{data.frame}

\item{labels}{list of column names and labels (see example)}
}
\value{
data.frame
}
\description{
Add labels to a data.frame
source: https://stackoverflow.com/questions/27347548/r-assign-variable-labels-of-data-frame-columns
}
\examples{
df <- data.frame(var_numeric = c(1, 2, 2, 3, NA, 4),
                 var_character = c("a", NA, "b", "c", "c", "d"),
                 stringsAsFactors = FALSE)
df <- SR_labeller_function(df,
                           c(var_numeric = "a",
                             var_character = "b"))
Hmisc::contents(df)

}
