% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_max_nchar_per_column.R
\name{SR_max_nchar_per_column}
\alias{SR_max_nchar_per_column}
\title{Determine the maximum nchar for each column of a data.frame}
\usage{
SR_max_nchar_per_column(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Determine the maximum nchar for each column of a data.frame.
}
\examples{
df <- data.frame(var_numeric = c(1, 2, 2, 3, NA, 44),
                 var_character = c("abc", NA, "bd", "cdefg", "c", "ef"),
                 stringsAsFactors = FALSE)
SR_max_nchar_per_column(df)

}
