% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_memory_usage.R
\name{SR_memory_usage}
\alias{SR_memory_usage}
\title{Shows memory usage}
\usage{
SR_memory_usage(
  pos = 1,
  pattern,
  order.by,
  decreasing = FALSE,
  head = FALSE,
  n = 5
)
}
\arguments{
\item{pos}{integer}

\item{pattern}{data.frame}

\item{order.by}{character}

\item{decreasing}{boolean}

\item{head}{boolean}

\item{n}{integer}
}
\value{
data.frame
}
\description{
Shows memory usage of objects in session/memory.
}
\examples{
SR_memory_usage()

}
