% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_modus.R
\name{SR_modus}
\alias{SR_modus}
\title{Calculate modus}
\usage{
SR_modus(x)
}
\arguments{
\item{x}{vector}
}
\value{
character
}
\description{
calculate most frequent value in a vector
}
\examples{
SR_modus(c("a", "a", "b", "c"))

}
