% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_mosaicplot.R
\name{SR_mosaicplot}
\alias{SR_mosaicplot}
\title{Mosaic plot}
\usage{
SR_mosaicplot(var1, var2)
}
\arguments{
\item{var1}{factor variable}

\item{var2}{factor variable}
}
\value{
ggplot-graphic
}
\description{
Plot a frequency count for example a confusion matrix (true y vs. predicted y).
}
