% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_NA_plot.R
\name{SR_NA_plot}
\alias{SR_NA_plot}
\title{NA plot}
\usage{
SR_NA_plot(
  df,
  save = FALSE,
  filename = "NA_Plot.png",
  path_output = path_output
)
}
\arguments{
\item{df}{data.frame}

\item{save}{Boolean}

\item{filename}{character, must end with "*png"}

\item{path_output}{character}
}
\value{
matrix, prints 2 plots
}
\description{
Plots (1) the occurence of NAs and (2) correlation of them in a data.frame.
}
\examples{
SR_NA_plot(df = data.frame(numeric = c(0, NA, 0.3, 0.5, NA, 1),
                           character = c("a", "b", NA, "d", NA, "f"),
                           integer = c(0L, NA, 3L, 5L, NA, 1L)))

}
