% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_omit_non_regular_values.R
\name{SR_omit_non_regular_values}
\alias{SR_omit_non_regular_values}
\title{Omits non regular values}
\usage{
SR_omit_non_regular_values(x)
}
\arguments{
\item{x}{a data.frame, matrix, vector or factor}
}
\value{
the same class as the input, but with omited rows of non regular values
}
\description{
Omits observations with values which are non regular (like: Inf, NA, NaN, -Inf)
}
\details{
Author: Rene Locher
Version: 2005-10-17
source: https://rdrr.io/cran/IDPmisc/src/R/NaRV.omit.R
}
\examples{
SR_omit_non_regular_values(as.vector(c(1, Inf, NaN, NA, -Inf)))
SR_omit_non_regular_values(as.vector(c("a", NaN, NA)))
SR_omit_non_regular_values(as.factor(c("a", Inf, NaN, NA, -Inf)))
SR_omit_non_regular_values(data.frame(a = c(1, Inf, NaN, NA, -Inf), b = 1:5))
SR_omit_non_regular_values(matrix(data = c(1, Inf, NaN, NA, -Inf, 1:5), ncol = 2))

}
