% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_open_folder_in_explorer.R
\name{SR_open_folder_in_explorer}
\alias{SR_open_folder_in_explorer}
\title{Open folder in Windows Explorer}
\usage{
SR_open_folder_in_explorer(path = getwd())
}
\arguments{
\item{path}{Path to an existing folder. Default is the working directory.
Windows paths with backslashes need a raw string, e.g.
\code{r"(C:\\temp)"}, otherwise R fails already when parsing.}
}
\value{
None
}
\description{
Opens a Windows Explorer window of a folder. Works only on Windows.
}
\examples{
\dontrun{
SR_open_folder_in_explorer("R")
SR_open_folder_in_explorer(r"(C:\home\sandbox\sandbox\Libraries)")
}

}
