% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_open_github.R
\name{SR_open_github}
\alias{SR_open_github}
\title{Open GitHub repository}
\usage{
SR_open_github()
}
\value{
None
}
\description{
Opens the GitHub repository of the current project in the browser.
}
\examples{
\dontrun{
SR_open_github()
}

}
