% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_plot_true_y_VS_predicted_y.R
\name{SR_plot_true_y_VS_predicted_y}
\alias{SR_plot_true_y_VS_predicted_y}
\title{Plot true y vs. predicted y}
\usage{
SR_plot_true_y_VS_predicted_y(
  true_y,
  predicted_y,
  path_output = NULL,
  save = F
)
}
\arguments{
\item{true_y}{numeric}

\item{predicted_y}{numeric}

\item{path_output}{character}

\item{save}{Boolean}
}
\value{
Boolean value TRUE or FALSE
}
\description{
e.g. plot a model prediction for y against true y values and calculates several precision metrics.
}
\examples{
data("mtcars")
SR_plot_true_y_VS_predicted_y(true_y = mtcars$hp,
                              predicted_y = mtcars$hp +
                                rnorm(nrow(mtcars), sd = 30))

}
