% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_plot_y_by_x.R
\name{SR_plot_y_by_x}
\alias{SR_plot_y_by_x}
\title{Plot y by x}
\usage{
SR_plot_y_by_x(df, x_name, y_name, path_output = path_output, save = F)
}
\arguments{
\item{df}{data.frame}

\item{x_name}{character}

\item{y_name}{character}

\item{path_output}{character}

\item{save}{Boolean}
}
\value{
Null, generates plot
}
\description{
Generates different univariate plots, depending if x/y are numeric, or factor/character
}
\examples{
data("mtcars")
SR_plot_y_by_x(mtcars,
               x = "hp", y = "mpg")
SR_plot_y_by_x(mtcars \%>\% mutate(hp = as.factor(hp)),
               x = "hp", y = "mpg")

}
