% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_remove_column_with_unique_value.R
\name{SR_remove_column_with_unique_value}
\alias{SR_remove_column_with_unique_value}
\title{Remove columns with unique value}
\usage{
SR_remove_column_with_unique_value(df, remove_na = FALSE, silent = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{remove_na}{Boolean: should NA values be removed before counting different values?}

\item{silent}{Boolean}
}
\value{
data.frame
}
\description{
Clean up data.frame by removing unnecessary columns, e.g. columns that have only 1 value.
}
\examples{
# removes the variable "remove" because all values are "1"
SR_remove_column_with_unique_value(df = data.frame(numeric = c(0, NA, 0.3, 0.5, NA, 1),
                                                   character = c("a", "b", NA, "d", NA, "f"),
                                                   integer = c(0L, NA, NA, NA, NA, NA),
                                                   remove = c(1, 1, 1, 1, 1, 1)))

# in addition removes variable "integer" because it only has the values "0" and "NA"
SR_remove_column_with_unique_value(df = data.frame(numeric = c(0, NA, 0.3, 0.5, NA, 1),
                                                   character = c("a", "b", NA, "d", NA, "f"),
                                                   integer = c(0L, NA, NA, NA, NA, NA),
                                                   remove = c(1, 1, 1, 1, 1, 1)),
                                   remove_na = TRUE)

}
