% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_replace_NA_modus.R
\name{SR_replace_NA_modus}
\alias{SR_replace_NA_modus}
\title{Replace NAs with modus}
\usage{
SR_replace_NA_modus(df, var = NULL, use_other_df = NULL)
}
\arguments{
\item{df}{data.frame}

\item{var}{character (optional)}

\item{use_other_df}{data.frame}
}
\value{
Boolean value TRUE or FALSE
}
\description{
Two possibilities:
1. If "var" is defined as a variable name (character):
All NA values of this variable will be replaced with modus of that variable.
2. If no "var" is defined:
All character and factor variables in the data.frame are examined.
In both cases a new variable will be created that retains information about the
replaced missing values.
If "use_other_df" is provided, then the modus is calculated on that data.frame
(for example calculated on training data and imputed on test data).
}
\examples{
SR_replace_NA_modus(df = data.frame(a = factor(c(1, 2, 2, 3, NA))),
                    var = "a")
SR_replace_NA_modus(df = data.frame(a = factor(c(1, 2, 2, 3, NA)),
                                    b = c("a", NA, "b", "c", "c")),
                    use_other_df = data.frame(a = factor(c(4, 5, 5, 6, 7)),
                                              b = c("e", "f", "X", "X", "h")))

}
