% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_replace_umlaute.R
\name{SR_replace_umlaute}
\alias{SR_replace_umlaute}
\title{Replace special characters in text}
\usage{
SR_replace_umlaute(df, cols = "")
}
\arguments{
\item{df}{data.frame}

\item{cols}{character (default = TRUE): select columns explicitly}
}
\value{
data.frame
}
\description{
Replaces the most common special characters.
}
\examples{
df <- data.frame(var_numeric = c(1, 2, 2, 3, NA, 4),
                 var_character = c("a", NA, "b", "c", "c", "d"),
                 var_factor = factor(c("a", "b", "c", "c", "d", NA)),
                 var_date = seq.Date(as.Date("2020-04-12"), by = "day", length.out = 6),
                 stringsAsFactors = FALSE)
SR_replace_umlaute(df)
rm(df)

}
