% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_rm_from_env.R
\name{SR_rm_from_env}
\alias{SR_rm_from_env}
\title{Remove objects from global environment by pattern}
\usage{
SR_rm_from_env(pattern)
}
\arguments{
\item{pattern}{character: regular expression passed to \code{ls()}}
}
\value{
None
}
\description{
Removes all objects whose names match a regular expression from the global
environment.
}
\examples{
\dontrun{
SR_rm_from_env(pattern = "^tmp_")
}

}
