% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_search_variables.R
\name{SR_search_variables}
\alias{SR_search_variables}
\title{Search variable(s) in a data.frame}
\usage{
SR_search_variables(df, var)
}
\arguments{
\item{df}{data.frame}

\item{var}{string}
}
\value{
Null.
}
\description{
Finds matching and partial matching variables and prints a summary of it/them.
}
