% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_show_project_folder_in_new_window.R
\name{SR_show_project_folder_in_new_window}
\alias{SR_show_project_folder_in_new_window}
\title{Opens the project folder in Windows Explorer}
\usage{
SR_show_project_folder_in_new_window()
}
\value{
NULL, opens Windows Explorer
}
\description{
Opens the project folder in Windows Explorer.
}
\examples{
SR_show_project_folder_in_new_window()

}
