% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_time_to_numeric.R
\name{SR_time_to_numeric}
\alias{SR_time_to_numeric}
\title{Calculate numeric value of a timestamp}
\usage{
SR_time_to_numeric(x)
}
\arguments{
\item{x}{object or variable in format "HH:MM:SS"}
}
\value{
numeric object or variable
}
\description{
calculate numeric value of a timestamp from an object in format "HH:MM:SS"
}
\examples{
SR_time_to_numeric("12:30:05")

}
