% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_trim_text.R
\name{SR_trim_text}
\alias{SR_trim_text}
\title{Remove leading or trailing whitespace from text}
\usage{
SR_trim_text(x)
}
\arguments{
\item{x}{text or character variable with text}
}
\value{
text or character variable with text
}
\description{
removes leading or trailing whitespace from text
}
\examples{
SR_trim_text(" abc ")

}
