% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_visualize_df.R
\name{SR_visualize_df}
\alias{SR_visualize_df}
\title{Visualize data.frame}
\usage{
SR_visualize_df(df, sample = 1000, no_graphs = 6, save = FALSE, df_name = "")
}
\arguments{
\item{df}{data.frame}

\item{sample}{integer}

\item{no_graphs}{integer}

\item{save}{Boolean}

\item{df_name}{character, must end with "*png"}
}
\value{
NULL, prints plots
}
\description{
Plots all variables of a data.frame in groups of their classes (numeric,
factor/character, date)
}
\examples{
SR_visualize_df(data.frame(var_numeric = c(1, 2, 2, 3, 3, 4),
                           var_character = c("a", NA, "b", "c", "c", "d"),
                           var_factor = factor(c("a", "b", "c", "c", "d", NA)),
                           var_date = seq.Date(Sys.Date(), by = "day", length.out = 6),
                           stringsAsFactors = FALSE))

}
