% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_mosaicplot.R
\name{SR_mosaicplot}
\alias{SR_mosaicplot}
\title{Mosaic plot (true y vs. predicted y)}
\usage{
SR_mosaicplot(df, cutoff = 0.5)
}
\arguments{
\item{df}{data.frame with columns "y_true" and "predicted_y"}

\item{cutoff}{numeric (default = 0.5): cutoff value for predicted_y to count as "1"}
}
\value{
ggplot-graphic
}
\description{
Plot a frequency count for a binary confusion matrix (true y vs. predicted y).
}
\examples{
SR_mosaicplot(df = data.frame(y_true = c(0, 0, 0, 1, 1, 1),
                              predicted_y = c(0.8, 0.1, 0.4, 0.2, 0.9, 0.8)),
              cutoff = 0.5)

}
