% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_univariate_analysis.R
\name{SR_univariate_analysis}
\alias{SR_univariate_analysis}
\title{Calculate univarate analysis}
\usage{
SR_univariate_analysis(df, y_name, path_output = path_output, save = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{y_name}{character}

\item{path_output}{character}

\item{save}{Boolean}
}
\value{
Boolean value TRUE or FALSE
}
\description{
Generates different univariate plots for all variables in a data.frame vs. y
}
\examples{
data("mtcars")
SR_univariate_analysis(mtcars \%>\%
                         dplyr::select(hp, cyl, mpg) \%>\%
                         dplyr::mutate(cyl = as.factor(cyl)),
                       y_name = "mpg", save = FALSE)

}
